backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
gunicorn==20.1.0
python-decouple==3.8
django-filter==23.1
django-cors-headers==3.14.0
python-dotenv
pytz==2020.1
sqlparse==0.3.1
asgiref==3.3.2
requests==2.26.0
django-colorfield==0.9.0
drf-extra-fields==3.7.0